refactor(web): replace useTimesheet with specialized hooks across timesheet components
- Updated multiple components to utilize the new `useTimesheetQuery`, `useDeleteTimesheet`, and `useUpdateTimesheet` hooks, enhancing code organization and maintainability.
- Removed direct usage of the old `useTimesheet` hook, streamlining the components and improving separation of concerns.
- Adjusted imports to reflect the new structure, ensuring better accessibility of utility functions from `timesheet-grouping`.
