+++ apps/web/core/components/timesheet/monthly-timesheet-calendar.tsx
@@ -1,6 +1,6 @@
 import React, { useMemo, useState, useCallback, useEffect } from 'react';
 import { format, addMonths, eachDayOfInterval, startOfMonth, endOfMonth, addDays, Locale, isLeapYear } from 'date-fns';
-import { GroupedTimesheet } from '@/core/hooks/activities/use-timesheet';
+import type { GroupedTimesheet } from '@/core/lib/helpers/timesheet-grouping';
 import { enGB } from 'date-fns/locale';
 import { cn } from '@/core/lib/helpers';
 import { formatDate } from '@/core/lib/helpers/index';
